pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "discord-py>=2.4.0",
    "openai>=1.55.3",
    "python-dotenv>=1.0.1",
]
